refactor: folder structure and files arranged in proper folder
