+++ src/data_store/mod.rs
@@ -1,2 +1,5 @@
+mod errors;
+mod implementations;
 pub mod store;
+mod types;
 mod value_entry;
